Move search logics to external hook
